requirements.txt
Kivy==2.0.0
kivymd==0.104.1
py-bcrypt==0.4
pycryptodome==3.10.1
